requirements.txt
Django==1.10.1
django-model-utils==2.6
djangorestframework==3.4.7
djangorestframework-csv==1.4.1
mysqlclient==1.3.7
six==1.10.0
xlrd==1.0.0
xlwt==1.2.0
PyMySQL==0.7.10
